requirements.txt
numpy
matplotlib
pandas
tqdm
requests
scikit-learn
scikit-surprise